スコア系float変換

+++ DoNotSit/Assets/Script/Enemy/Direction_Collision/Direction_Forward.cs
@@ -14,7 +14,7 @@ public class Direction_Forward : MonoBehaviour
     float cnt;
     Renderer render;
     PlayerControl player;
-    public int score;
+    public float score;
 
     void Start()
     {
